breaking app.go into loader and transformer, to be continue
